refactor: extract status and event loop methods to PollingOperations
Move _set_api_status, _set_fcm_status, api_status, fcm_status,
is_fcm_connected, consecutive_timeouts, last_poll_result, _is_on_hass_loop,
_run_on_hass_loop (~90 lines) from main.py to polling.py.
